crt: move git_utils from crt to cbscommon.git
* what:
  * Move the SHA type definition from crt.crtlib.git_utils to a new
    module cbscommon.git._types.
  * Re-export SHA in cbscommon.git.__init__.
  * Update all modules in the crt project to import SHA from its
    new location in cbscommon.git.
  * Add comments explaining the migration in all affected files.
  * Create `cbscommon.exceptions` with `CBSCommonError` base class.
  * Create `cbscommon.git._exceptions` containing all git-related exceptions
    previously in `crt.crtlib.git_utils`.
  * Update `GitError` to inherit from `CBSCommonError`.
  * Update all `crt` modules to import git exceptions from `cbscommon.git`.
  * Re-export all git exceptions in `cbscommon.git.__init__`.
  * Renamed and moved crt/src/crt/crtlib/git_utils.py to
    cbscommon/src/cbscommon/git/_cmds.py.
  * Updated cbscommon/src/cbscommon/git/__init__.py to export all git
    utility functions (e.g., git_push, git_checkout_ref, etc.)
    via __all__.
  * Refactored imports across several crt modules (manifest.py,
    patch.py, release.py, apply.py, etc.) to pull git functions directly
    from cbscommon.git.
  * Updated the cmds.py logging to use a standalone cbscommon.git logger
    instead of a child of the crt logger.
  * Switched to SPDX license identifiers in the moved file.
  * Removed the legacy if __name__ == "__main__": test block from the
    git utilities.

* why:
  * To centralize common git types in cbscommon, enabling their use
    across different projects in the CBS ecosystem without cyclic
    dependencies.
  * To share git-related exceptions across different projects in the CBS
    ecosystem and establish a common error hierarchy.
  * Centralization: Centralizes shared Git logic into a common library
    to allow other services or packages to reuse these utilities without
    a dependency on crt.
  * Clean Architecture: Separates low-level Git wrapper operations from
    high-level crt command logic.
  * Maintainability: Standardizes how Git exceptions and functions are
    accessed across the codebase, reducing duplicate logic and import
    complexity.
