+++ cbscommon/src/cbscommon/git/_cmds.py
@@ -1,155 +1,35 @@
-# crt - utils
-# Copyright (C) 2025  Clyso GmbH
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2026 Clyso GmbH
+
 
 import logging
 import re
-import sys
 import tempfile
+from collections.abc import Callable
 from pathlib import Path
-from typing import cast, override
+from typing import cast
 
 import git
 
-from crt.crtlib.errors import CRTError
-from crt.crtlib.logger import logger as parent_logger
+from cbscommon.git._exceptions import (
+    GitAMApplyError,
+    GitCherryPickConflictError,
+    GitCherryPickError,
+    GitCreateHeadExistsError,
+    GitEmptyPatchDiffError,
+    GitError,
+    GitFetchError,
+    GitFetchHeadNotFoundError,
+    GitHeadNotFoundError,
+    GitIsTagError,
+    GitMissingBranchError,
+    GitMissingRemoteError,
+    GitPatchDiffError,
+    GitPushError,
+)
+from cbscommon.git._types import SHA
 
-logger = parent_logger.getChild("git")
-
-
-SHA = str
-
-
-class GitError(CRTError):
-    @override
-    def __str__(self) -> str:
-        return "git error" + (f": {self.msg}" if self.msg else "")
-
-
-class GitIsTagError(GitError):
-    def __init__(self, tag: str) -> None:
-        super().__init__(msg=tag)
-
-    @override
-    def __str__(self) -> str:
-        return self.with_maybe_msg("found unexpected tag")
-
-
-class GitPatchDiffError(GitError):
-    @override
-    def __str__(self) -> str:
-        return "patches diff error" + (f": {self.msg}" if self.msg else "")
-
-
-class GitEmptyPatchDiffError(GitPatchDiffError):
-    pass
-
-
-class GitCherryPickError(GitError):
-    @override
-    def __str__(self) -> str:
-        return "cherry-pick error" + (f": {self.msg}" if self.msg else "")
-
-
-class GitCherryPickConflictError(GitCherryPickError):
-    sha: SHA
-    conflicts: list[str]
-
-    def __init__(self, sha: SHA, files: list[str] | None = None) -> None:
-        super().__init__(msg=f"conflict occurred on sha '{sha}'")
-        self.sha = sha
-        self.conflicts = files if files else []
-
-
-class GitAMApplyError(GitError):
-    pass
-
-
-class GitMissingRemoteError(GitError):
-    remote_name: str
-
-    def __init__(self, remote_name: str) -> None:
-        super().__init__()
-        self.remote_name = remote_name
-
-    @override
-    def __str__(self) -> str:
-        return f"missing remote '{self.remote_name}'"
-
-
-class GitMissingBranchError(GitError):
-    def __init__(self, branch_name: str) -> None:
-        super().__init__(msg=branch_name)
-
-    @override
-    def __str__(self) -> str:
-        return self.with_maybe_msg("missing branch")
-
-
-class GitCreateHeadExistsError(GitError):
-    def __init__(self, name: str) -> None:
-        super().__init__(msg=name)
-
-    @override
-    def __str__(self) -> str:
-        return self.with_maybe_msg("head already exists")
-
-
-class GitHeadNotFoundError(GitError):
-    def __init__(self, name: str) -> None:
-        super().__init__(msg=name)
-
-    @override
-    def __str__(self) -> str:
-        return self.with_maybe_msg("head not found")
-
-
-class GitFetchError(GitError):
-    remote_name: str
-    from_ref: str
-    to_ref: str
-
-    def __init__(self, remote_name: str, from_ref: str, to_ref: str) -> None:
-        super().__init__()
-        self.remote_name = remote_name
-        self.from_ref = from_ref
-        self.to_ref = to_ref
-
-    @override
-    def __str__(self) -> str:
-        return (
-            f"failed fetching from remote '{self.remote_name}': "
-            + f"from '{self.from_ref}' to '{self.to_ref}'"
-        )
-
-
-class GitFetchHeadNotFoundError(GitFetchError):
-    def __init__(self, remote_name: str, head: str) -> None:
-        super().__init__(remote_name, head, "")
-
-    @override
-    def __str__(self) -> str:
-        return f"head '{self.from_ref}' not found in remote '{self.remote_name}'"
-
-
-class GitPushError(GitError):
-    def __init__(self, branch: str, dst_branch: str, remote_name: str) -> None:
-        super().__init__(
-            msg=f"from branch '{branch}' to '{dst_branch}' on remote '{remote_name}'"
-        )
-
-    @override
-    def __str__(self) -> str:
-        return self.with_maybe_msg("unable to push")
+logger = logging.getLogger("cbscommon.git")
 
 
 def git_check_patches_diff(
@@ -814,56 +694,10 @@ def git_format_patch(repo_path: Path, rev: SHA, *, base_rev: SHA | None = None) 
 def git_tag_exists_in_remote(repo_path: Path, remote_name: str, tag_name: str) -> bool:
     try:
         repo = git.Repo(repo_path)
-        raw_tag: str = repo.git.ls_remote(
-            "--tags", remote_name, f"refs/tags/{tag_name}"
-        )
+        ls_remote = cast(Callable[..., str], repo.git.ls_remote)
+        raw_tag = ls_remote("--tags", remote_name, f"refs/tags/{tag_name}")
         return bool(raw_tag.strip())
     except git.CommandError as e:
         msg = f"unable to execute git ls-remote --tags {remote_name} refs/tags/{tag_name}: {e}"
         logger.error(msg)
         raise GitError(msg) from None
-
-
-if __name__ == "__main__":
-    if len(sys.argv) < 2:
-        print("error: missing repo path argument")
-        sys.exit(1)
-
-    logger.setLevel(logging.DEBUG)
-
-    repo_path = Path(sys.argv[1])
-
-    print("checkout refs")
-    try:
-        git_checkout_ref(repo_path, "foobar")
-    except Exception as e:
-        print(f"error getting 'foobar': {e}")
-
-    try:
-        git_checkout_ref(
-            repo_path, "main", update_from_remote=True, remote_name="clyso/ceph"
-        )
-    except Exception as e:
-        print(f"error getting 'foobar': {e}")
-
-    try:
-        git_checkout_ref(
-            repo_path,
-            "tentacle",
-            update_from_remote=True,
-            remote_name="clyso/ceph",
-        )
-    except Exception as e:
-        print(f"error getting 'tentacle': {e}")
-
-    try:
-        git_checkout_ref(
-            repo_path,
-            "v18.2.7",
-            to_branch="test-v18.2.7",
-            remote_name="ceph/ceph",
-            update_from_remote=True,
-            fetch_if_not_exists=True,
-        )
-    except Exception as e:
-        print(f"error checking out 'v18.2.7' to 'test-v18.2.7': {e}")
