+++ cbscommon/src/cbscommon/git/_types.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2026 Clyso GmbH
+
+
+# SHA type was moved from crt.crtlib.git_utils
+SHA = str
