+++ crt/src/crt/cmds/patch.py
@@ -18,12 +18,16 @@ from collections.abc import Callable
 from pathlib import Path
 
 import click
+from cbscommon.git import (
+    GitError,
+    git_prepare_remote,
+    git_revparse,
+)  # Git exceptions and functions are now imported from cbscommon.git
 
 from crt.cmds import Ctx, pass_ctx, perror, psuccess, pwarn, with_patches_repo_path
 from crt.cmds import logger as parent_logger
 from crt.crtlib.apply import ApplyError, patches_apply_to_manifest
 from crt.crtlib.errors.manifest import MalformedManifestError, NoSuchManifestError
-from crt.crtlib.git_utils import GitError, git_prepare_remote, git_revparse
 from crt.crtlib.manifest import load_manifest_by_name_or_uuid, store_manifest
 from crt.crtlib.patch import (
     PatchError,
