+++ crt/src/crt/cmds/patchset.py
@@ -23,6 +23,15 @@ from typing import cast
 import click
 import pydantic
 import rich.box
+from cbscommon.git import (
+    SHA,
+    GitError,
+    git_branch_delete,
+    git_branch_from,
+    git_get_patch_sha_title,
+    git_patches_in_interval,
+    git_prepare_remote,
+)  # Git types, exceptions and functions are now imported from cbscommon.git
 from rich.console import Group, RenderableType
 from rich.padding import Padding
 from rich.table import Table
@@ -41,15 +50,6 @@ from crt.cmds import logger as parent_logger
 from crt.cmds._common import CRTProgress
 from crt.crtlib.errors.patchset import (
     MalformedPatchSetError,
-)
-from crt.crtlib.git_utils import (
-    SHA,
-    GitError,
-    git_branch_delete,
-    git_branch_from,
-    git_get_patch_sha_title,
-    git_patches_in_interval,
-    git_prepare_remote,
 )
 from crt.crtlib.models.common import (
     AuthorData,
