+++ crt/src/crt/crtlib/errors/manifest.py
@@ -14,8 +14,9 @@
 import uuid
 from typing import override
 
+from cbscommon.git import SHA  # SHA is now imported from cbscommon.git
+
 from crt.crtlib.errors import CRTError
-from crt.crtlib.git_utils import SHA
 from crt.crtlib.models.common import AuthorData
