+++ crt/src/crt/crtlib/models/patch.py
@@ -16,8 +16,8 @@ from datetime import datetime as dt
 from typing import override
 
 import pydantic
+from cbscommon.git import SHA  # SHA is now imported from cbscommon.git
 
-from crt.crtlib.git_utils import SHA
 from crt.crtlib.models.common import (
     AuthorData,
     ManifestPatchEntry,
