+++ crt/src/crt/crtlib/models/patchset.py
@@ -17,9 +17,9 @@ from datetime import datetime as dt
 from typing import Annotated, Any, override
 
 import pydantic
+from cbscommon.git import SHA  # SHA is now imported from cbscommon.git
 
 from crt.crtlib.errors.patchset import EmptyPatchSetError
-from crt.crtlib.git_utils import SHA
 from crt.crtlib.models.common import (
     AuthorData,
     ManifestPatchEntry,
